# twep-backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TWEP",
    "description": "TWEP API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string"
          },
          "wallet": {
            "type": "number",
            "format": "float"
          },
          "tickets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ticket"
            }
          }
        },
        "required": [
          "email",
          "password"
        ],
        "example": {
          "id": 1,
          "email": "user@example.com",
          "password": "Password123",
          "role": "user",
          "wallet": 100
        }
      },
      "UserLogin": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "BikeType": {
        "type": "string",
        "enum": [
          "Mountain Bike",
          "City Bike",
          "Electric Bike",
          "Children Bike"
        ]
      },
      "PurchaseTicket": {
        "type": "object",
        "properties": {
          "bikeType": {
            "$ref": "#/components/schemas/BikeType"
          },
          "station": {
            "type": "string"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "immediateRenting": {
            "type": "boolean"
          },
          "reservedStation": {
            "type": "string"
          }
        },
        "required": [
          "bikeType",
          "station",
          "purchaseDate",
          "immediateRenting"
        ],
        "example": {
          "bikeType": "Mountain Bike",
          "station": "Station A",
          "purchaseDate": "2024-01-10T08:00:00Z",
          "immediateRenting": true,
          "reservedStation": "Station B"
        }
      },
      "Ticket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "bikeType": {
            "$ref": "#/components/schemas/BikeType"
          },
          "station": {
            "type": "string"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "immediateRenting": {
            "type": "boolean"
          },
          "reservedStation": {
            "type": "string"
          }
        },
        "required": [
          "bikeType",
          "station",
          "purchaseDate",
          "immediateRenting"
        ],
        "example": {
          "id": 1,
          "bikeType": "Mountain Bike",
          "station": "Station A",
          "purchaseDate": "2024-01-10T08:00:00Z",
          "immediateRenting": true,
          "reservedStation": "Station B"
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "example": {
          "error": "Some error message"
        }
      },
...
    }
  },
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "user"
        ],
        "requestBody": {
          "description": "User registration details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "example": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "example": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "User login",
        "tags": [
          "user"
        ],
        "requestBody": {
          "description": "User login details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid credentials"
                }
              }
            }
          }
        }
      }
    },
    "/users/profile": {
      "get": {
        "summary": "Get user profile",
        "tags": [
          "user"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "example": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "User not found"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/account": {
      "get": {
        "summary": "Get user account details",
        "tags": [
          "user"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "User not found"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/account/add-money": {
      "post": {
        "summary": "Add money to user's wallet",
        "tags": [
          "user"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the user"
          }
        ],
        "requestBody": {
          "description": "Amount to add to the wallet",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "format": "float"
                  }
                },
                "required": [
                  "amount"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Wallet updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "email": "user@example.com",
                  "wallet": 150
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid input data"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "User not found"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/account/tickets": {
      "get": {
        "summary": "Get user's ticket history",
        "tags": [
          "user"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "tickets": [
                    {
                      "id": 1,
                      "bikeType": "Mountain Bike",
                      "station": "Station A",
                      "purchaseDate": "2024-01-10T08:00:00Z"
                    },
                    {
                      "id": 2,
                      "bikeType": "City Bike",
                      "station": "Station B",
                      "purchaseDate": "2024-01-15T10:30:00Z"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "User not found"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Purchase a new ticket",
        "tags": [
          "user"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the user"
          }
        ],
        "requestBody": {
          "description": "Ticket details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseTicket"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket purchased successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "bikeType": "Mountain Bike",
                  "station": "Station A",
                  "purchaseDate": "2024-01-10T08:00:00Z",
                  "immediateRenting": true,
                  "reservedStation": "Station B"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Insufficient funds in the wallet for ticket purchase"
                }
              }
            }
          }
        }
      }
    },
    "/users/stations": {
      "get": {
        "summary": "Get list of bike stations",
        "tags": [
          "station"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "stations": [
                    {
                      "id": "uuid",
                      "name": "Bike Station 1",
                      "location": {
                        "latitude": 40.7128,
                        "longitude": -74.006
                      },
                      "parkingPlaces": [
                        {
                          "id": "uuid",
                          "bikeCategories": [
                            "city_bike",
                            "mountain_bike"
                          ]
                        }
                      ],
                      "operational": true
                    },
                    {
                      "id": "uuid",
                      "name": "Bike Station 2",
                      "location": {
                        "latitude": 40.73,
                        "longitude": -73.995
                      },
                      "parkingPlaces": [
                        {
                          "id": "uuid",
                          "bikeCategories": [
                            "electric_bike",
                            "children_bike"
                          ]
                        }
                      ],
                      "operational": true
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Stations not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Stations not found"
                }
              }
            }
          }
        }
      }
    },
    "/users/stations/{stationId}/bikes": {
      "get": {
        "summary": "Get list of bikes available at a specific station",
        "tags": [
          "bike"
        ],
        "parameters": [
          {
            "name": "stationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The ID of the station"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "bikes": [
                    {
                      "id": "uuid",
                      "status": "available",
                      "category": "Electric Bike",
                      "model": "Model X",
                      "description": "Electric Bike, powerful motor",
                      "wheelSize": 26.5,
                      "extraFeatures": [
                        "Front suspension",
                        "GPS device"
                      ],
                      "parkingPlaceId": "c2c6b68e-759c-496a-a860-ee27c8874884"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Station not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Station not found"
                }
              }
            }
          }
        }
      }
    },
    "/users/bikes": {
      "get": {
        "summary": "Get list of all bikes available",
        "tags": [
          "bike"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "bikes": [
                    {
                      "id": "48661dea-6625-4376-a593-85ecd8c711ed",
                      "status": "available",
                      "categoryId": "3a4dfc98-79da-4a7b-875d-5122008119d5",
                      "category": "Mountain Bikes",
                      "modelId": "2e5ae85b-77f2-4d28-bace-fbcd467bcb9a",
                      "model": "KTM 1.0",
                      "description": "KTM's finest new addition. Can go up any mountain and down any valley",
                      "wheelSize": 25,
                      "extraFeatures": [
                        "Big wheels",
                        "Comfortable seat"
                      ],
                      "parkingPlaceId": "14891dac-c49e-4dcb-95ef-c994891569af",
                      "station": {
                        "id": "6361af2e-25ee-469e-9d8f-2495b63f0270",
                        "name": "Alpenstraße",
                        "location": {
                          "latitude": 47.77,
                          "longitude": 13.07029
                        },
                        "operational": true
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Station not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Station not found"
                }
              }
            }
          }
        }
      }
    },
    "/users/bike-categories": {
      "get": {
        "summary": "Get a list of all bike categories",
        "tags": [
          "bike"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "bikeCategories": [
                    "mountain_bike",
                    "city_bike",
                    "electric_bike",
                    "children_bike"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Bike categories not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Bike categories not found"
                }
              }
            }
          }
        }
      }
    },
    "/users/bike-models": {
      "get": {
        "summary": "Get a list of all bike models",
        "tags": [
          "bike"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "bikeModels": [
                    {
                      "name": "Mountain Bike XL",
                      "description": "A robust mountain bike for off-road adventures",
                      "wheelSize": 29.5,
                      "extraFeatures": [
                        "Shimano XT gears",
                        "Front suspension with 120mm travel"
                      ]
                    },
                    {
                      "name": "City Cruiser",
                      "description": "A comfortable city bike for daily commuting",
                      "wheelSize": 28,
                      "extraFeatures": [
                        "Built-in rear rack",
                        "Internal hub gears"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Bike models not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Bike models not found"
                }
              }
            }
          }
        }
      }
    },
    "/users/stations/{id}": {
      "get": {
        "summary": "Get details of a specific bike station",
        "tags": [
          "station"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The ID of the station"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "example": {
                  "id": "uuid",
                  "name": "Bike Station 1",
                  "location": {
                    "latitude": 40.7128,
                    "longitude": -74.006
                  },
                  "parkingPlaces": [
                    {
                      "id": "uuid",
                      "bikeCategories": [
                        "city_bike",
                        "mountain_bike"
                      ]
                    }
                  ],
                  "operational": true
                }
              }
            }
          },
          "404": {
            "description": "Station not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Station not found"
                }
              }
            }
          }
        }
      }
    },
...
  }
}